# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Schedule",
        "description": "Documentation de l'API Schedule",
        "contact": {
            "email": "wwwmbassiloic@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "API "
        }
    ],
    "paths": {
...
        "/user-group/": {
            "get": {
                "tags": [
                    "userGroup"
                ],
                "summary": "Get list of userGroup",
                "description": "Returns list of userGroup",
                "operationId": "getuserGroupList",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/user-group/{id}": {
            "get": {
                "tags": [
                    "userGroup"
                ],
                "summary": "Get userGroup information",
                "description": "Returns userGroup data",
                "operationId": "222f282224f142d912aa25022b66a93c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "userGroup id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/user-group/delete/{id}": {
            "delete": {
                "tags": [
                    "userGroup"
                ],
                "summary": "Delete existing userGroup",
                "description": "Deletes a record and returns no content",
                "operationId": "ca1a0b0b15056ea948efc6a7fcfc0b1f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "userGroup id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/user-group/save": {
            "post": {
                "tags": [
                    "userGroup"
                ],
                "summary": "Store new userGroup",
                "description": "Returns userGroup data",
                "operationId": "c3384034e41be7a25979495eaf1e0f24",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/user-group/update/{id}": {
            "put": {
                "tags": [
                    "userGroup"
                ],
                "summary": "Update existing userGroup",
                "description": "Returns updated userGroup data",
                "operationId": "c0d37722dea59c600cdfb88c6cab51fc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "userGroup id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "202": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        }
    },
...
}